+++ frontend/src/types/common.d.ts
@@ -1,0 +1,1 @@
+type Setter<T> = React.Dispatch<React.SetStateAction<T>>
